Enhanced inheritance between layouts.
Each layout may specify a custom layout used for page layout inheritance.
Eg. the layout employed in the root page can now specify a different layout
for inner pages. This simplifies writing the layouts, because it is no longe
needed to check if the rendered page is the root (using if_parent, for
instance). The default behavior is to let inner pages inherit the same layou
their ancestors.